third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/encoding-send-latin-1.html

<meta charset="utf-8">
<script src="../../js-test-resources/js-test.js"></script>
<script>
description("Test that Latin-1 characters are ent to the server correctly.");

var tests = [
    "ASCII",
    "Latin-1: \xC5",
    "Fake UTF-8: \xC3\x85",
    "UTF-16: \u6587",
    "32-bit Character: \uD83D\uDCA9",
];

for (var i = 0; i < tests.length; ++i) {
    var xhr = new XMLHttpRequest();
    xhr.open("POST", "/xmlhttprequest/resources/post-echo-as-utf-8.cgi", false);
    xhr.send(tests[i]);
    shouldBe("xhr.responseText", "tests[" + i + "]");
}
</script>
