third_party/blink/web_tests/http/tests/xmlhttprequest/encoding-send-latin-1-expected.txt

Test that Latin-1 characters are ent to the server correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseText is tests[0]
PASS xhr.responseText is tests[1]
PASS xhr.responseText is tests[2]
PASS xhr.responseText is tests[3]
PASS xhr.responseText is tests[4]
PASS successfullyParsed is true

TEST COMPLETE
